Move OCA.Apps to the server bundle
