+++ core/js/core.json
@@ -12,7 +12,6 @@
 		"oc-backbone.js",
 		"oc-backbone-webdav.js",
 		"l10n.js",
-		"apps.js",
 		"share.js",
 		"sharetemplates.js",
 		"sharesocialmanager.js",
